requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
bcrypt==4.3.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.3
ecdsa==0.19.1
fastapi==0.109.2
greenlet==3.2.2
h11==0.16.0
idna==3.10
jwt==1.3.1
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.36.3
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.27.1
